% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{effective_dose}
\alias{effective_dose}
\title{Calculate an effective dose}
\usage{
effective_dose(object, probability = 0.5)
}
\arguments{
\item{object}{A fitted \code{qdr_fit} object.}

\item{probability}{Response probabilities strictly between zero and one.}
}
\value{
A numeric vector of doses.
}
\description{
Calculate an effective dose
}
