% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{glance.qdr_fit}
\alias{glance.qdr_fit}
\title{Glance at a dose-response fit}
\usage{
\method{glance}{qdr_fit}(x, ...)
}
\arguments{
\item{x}{A \code{qdr_fit} object.}

\item{...}{Unused.}
}
\value{
A one-row model summary tibble.
}
\description{
Glance at a dose-response fit
}
