% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{qrmavfar-package}
\alias{qrmavfar}
\alias{qrmavfar-package}
\title{qrmavfar: microbial dose-response model fitting}
\description{
Fits exponential and approximate beta-Poisson dose-response models to grouped binomial microbial response data. Provides trend tests, model diagnostics, bootstrap uncertainty estimates, effective doses, and publication-ready dose-response plots.
}
\author{
\strong{Maintainer}: Sean Thimons \email{thimons.sean@gmail.com}

Authors:
\itemize{
  \item Sean Thimons \email{thimons.sean@gmail.com}
}

Other contributors:
\itemize{
  \item Mark H. Weir [contributor]
  \item Timothy A. Bartrand [contributor]
}

}
\keyword{internal}
