% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{singlehit-package}
\alias{singlehit}
\alias{singlehit-package}
\title{singlehit: microbial dose-response model fitting}
\description{
Fits exponential and approximate beta-Poisson dose-response models to grouped binomial microbial response data. Provides trend tests, model diagnostics, bootstrap uncertainty estimates, effective doses, and publication-ready dose-response plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://seanthimons.github.io/singlehit/}
  \item \url{https://github.com/seanthimons/singlehit}
  \item Report bugs at \url{https://github.com/seanthimons/singlehit/issues}
}

}
\author{
\strong{Maintainer}: Sean Thimons \email{thimons.sean@gmail.com}

Authors:
\itemize{
  \item Sean Thimons \email{thimons.sean@gmail.com}
}

Other contributors:
\itemize{
  \item Mark H. Weir [contributor]
  \item Timothy A. Bartrand [contributor]
}

}
\keyword{internal}
